More tweaks and adjustments.
This commit makes changes to the display_message and spoken_message attributes where \n line breaks embedded in the alert feed cause the sentences of the UI notifications to look broken on mobile browsers. Single line breaks in the description and instruction text parts of the display_message and spoken_message are now replaced with single spaces and double line breaks are left in to preserve paragraph break points. I also changed the weather_alerts_are_active sensor states from on/off to Yes/No.