More updates and reverted some previous changes
I have reverted some changes that allowed multiple zones to be specified in the 'zone' config. Added a new config option 'type' that must be set to 'zone' or 'county' depending on which zone ID number you want to use for your alerts. Zone/County zone IDs should be specified as one, two, or three digit numbers only and not as the full six character feed ID. JSON alert keys that don't exist in the alert feed will be set to 'null'. If the alert feed times out or the server returns anything other than HTTP Status Code 200, the weatheralerts sensor will be set to 'unavailable' but will retain any alerts that were already there.