Updates and additions
Changes to the alert sensors to handle the condition of the main weatheralerts sensor having a state of 'unavailable' in the event the component can't download the alert feed. An addition of weatheralerts_1_alert_#_most_recent_active_alert template sensors. These will hold the current alert event if there is an active alert or it will hold the previously active alert event. On restart of Home Assistant, the state of these sensors will be 'unavailable' until an alert is available. Because new alerts push older alerts down the list, weatheralerts_1_alert_1_most_recent_active_alert is likely going to be the most useful of the 5 sensors when viewing state history to see previous alert events. These most_recent_active_alert sensors will likely change and evolve to include alert attributes in the future.