Initial commit - for use with second alert zone
If you want alerts for a second zone, this yaml is the same as weather_alerts.yaml but has all of the necessary changes to the sensors, automations, and script. Just change the state, zone, and type in the weatheralerts platform config near the top of the yaml (below the initial comments) to the appropriate values for your second zone. Follow instructions in the comments of the yaml to use this yaml package as-is (except for the weatheralerts platform config) with no template sensor, automation, or script changes needed.