Final updates to sensor.py for this pull request
I removed the 'type' config option and added a 'county' config option. The required 'zone' config option should now only contain the zone ID number and the optional 'county' config option should only contain a county ID number. Configuring both 'zone' and 'county' is highly recommended for increased reliability of alerts. This is a non-breaking change for the original weatheralerts component, but it is a breaking change if you installed any earlier sensor.py commits from this pull request.

I also added a different test to check the validity of zone and county IDs and log the appropriate errors to the Home Assistant log file. The previous test only worked in a relatively small percentage of cases and returned a valid json with no errors even though an invalid ID number was used.