Add county ID, additional alert properties, and PlatformNotReady (#29)
No breaking changes. If you have a weatheralerts platform configuration already set, it should continue to work. However, for alert reliability, it is highly recommended that you add the new 'county' platform config option to specify your county ID number (it is not likely going to be the same number as the zone ID number). Check the README.md documentation for details.

Added 'county' platform config option to specify an optional county ID number. 'zone' should be set to a proper zone ID number (and is required), the county ID number should then be specified with the 'county' option.

Additional alert properties have been added along with proper handling in the event an attribute is not included in the json alert feed. Feel free to request additional alert properties by submitting an issue here on github (https://github.com/custom-components/weatheralerts/issues).

The main platform sensor entity state will contain the number of active alerts, unless the sensor fails to connect with weather.gov to retrieve alerts. If the sensor cannot connect to weather.gov to retrieve the json alert feed, the sensor state will be set to 'unavailable'. The yaml package(s) with this pull request includes template sensors that will retain active alerts and the active alert count (for a short period), even when the main sensor becomes 'unavailable'.

PlatformNotReady is added to allow the platform to continue to retry the initialization if the weather.gov site is too slow to respond or experiencing other issues.

Added additional checks to verify the specified zone and county IDs are valid. 

Included a set of comprehensive yaml packages that include tracking for up to 5 active weather alerts and automations and script to handle UI notifications.

Started update of README.md documentation.