Add endsExpires attribute
This new 'endsExpires' attribute will be equal to the 'expires' attribute if the 'ends' attribute is null, otherwise it will be equal to the 'ends' attribute. If the weather.gov alerts API has an extended outage, this addition will make it easy to turn an alert off and stop counting it after the alert expires or ends and the integration is stuck with stale alerts due to not being able to download from the alerts feed.