Add three more alert properties
This adds three more alert properties from the weather.gov alert feed: hailSize, windGust, and waterspoutDetection. hailSize is a number (such as '0.25' or '1') and represents hail size in inches. windGust is a string that includes the wind gust speed and the unit of measurement (such as '60 MPH') in the string. waterspoutDetection is a string (such as 'POSSIBLE' or 'OBSERVED') and appears to be included as needed with marine warnings for marine zones.